fix(rust): `ProjectDatabaseDocumentRunQueryCall` return type
The Firestore discovery document claims that `runQuery()` returns a
`RunQueryResponse` object when in fact it returns an array of them.

Wrap `google_firestore1_beta1::ProjectDatabaseDocumentRunQueryCall::doit()`s
return type in a `Vec` to reflect this reality.

NB: this fix will be erased on the next re-generation.
TODO: find a way to fix this permanently.